Extradimensional Pursuit Update
Updated Launch Throw and Player Movement so the launch grab sends the arms to and from the player once, while not moving the arms when:
a) The grab button is held OR
b) The grab is in progress.

Added a Hitboxed Enemy variant with a 3D Rigidbody, and a script to match.