THE ONE THROW Update
THE ONE THROW IS REAAAAAAAL

Added 2D sprites package

Added scripts to make throwing functional

Hitting an Enemy with the grab results in that enemy being destroyed, the player's arms instantly returning to them, and a prefab being selected. This also allows the player to "throw" the enemy by pressing the grab button again.

After throwing an enemy, the grab will work as normal.

Replaced the hand graphics with hand sprites.